Moved and renamed some ui files
