+++ src/phases/learn-move-phase.ts
@@ -10,7 +10,7 @@ import { UiMode } from "#enums/ui-mode";
 import type { Pokemon } from "#field/pokemon";
 import type { Move } from "#moves/move";
 import { PlayerPartyMemberPokemonPhase } from "#phases/player-party-member-pokemon-phase";
-import { EvolutionSceneHandler } from "#ui/handlers/evolution-scene-handler";
+import { EvolutionSceneUiHandler } from "#ui/handlers/evolution-scene-ui-handler";
 import { SummaryUiMode } from "#ui/handlers/summary-ui-handler";
 import i18next from "i18next";
 
@@ -47,7 +47,7 @@ export class LearnMovePhase extends PlayerPartyMemberPokemonPhase {
     }
 
     this.messageMode =
-      globalScene.ui.getHandler() instanceof EvolutionSceneHandler ? UiMode.EVOLUTION_SCENE : UiMode.MESSAGE;
+      globalScene.ui.getHandler() instanceof EvolutionSceneUiHandler ? UiMode.EVOLUTION_SCENE : UiMode.MESSAGE;
     globalScene.ui.setMode(this.messageMode);
     // If the Pokemon has less than 4 moves, the new move is added to the largest empty moveset index
     // If it has 4 moves, the phase then checks if the player wants to replace the move itself.
