+++ src/ui/utils/pokemon-icon-anim-handler.ts
